Crate kywy

Modules

battery

Battery monitor for the Koinslot device. With battery icon drawing capabilities.

button_async

src/buttons.rs Button event system using a shared Channel for press/release detection.

button_poll

src/button_poll.rs Button polling library for Kywy board (direct level check, no async) Returns true (Pressed) or false (Released)

display

Kywy Display Driver

macros

sdcard

SD-Card reads and writes using async SPI bus does not apear to be currently supported.

Macros

kywy_battery_from

kywy_button_async_from

kywy_button_poll_from

kywy_display_from

kywy_spi_from